% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\docType{data}
\name{core10.dat}
\alias{core10.dat}
\title{The core10.dat database}
\source{
\url{https://github.com/MarcNeveu/IcyDwarf/blob/master/IcyDwarf/PHREEQC-3.1.2/core10.dat}
}
\usage{
core10.dat  # phrLoadDatabaseString(core10.dat)
}
\description{
core10.dat is a database derived from phreeqc.dat and
llnl.dat, with careful data checking, use of SUPCRT for temperature
dependence, and addition of molar volumes for pressure dependence.
Contributed by Marc Neveu. The database has been reformatted for use by
\code{\link{phrLoadDatabaseString}}.
}
\references{
\doi{10.1016/j.gca.2017.06.023}
}
\seealso{
Other Databases: 
\code{\link{Amm.dat}},
\code{\link{ColdChem.dat}},
\code{\link{Tipping_Hurley.dat}},
\code{\link{ex15.dat}},
\code{\link{frezchem.dat}},
\code{\link{iso.dat}},
\code{\link{llnl.dat}},
\code{\link{minteq.dat}},
\code{\link{minteq.v4.dat}},
\code{\link{phreeqc.dat}},
\code{\link{pitzer.dat}},
\code{\link{sit.dat}},
\code{\link{wateq4f.dat}}
}
\concept{Databases}
\keyword{dataset}
