% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\docType{data}
\name{frezchem.dat}
\alias{frezchem.dat}
\title{The frezchem.dat database}
\usage{
frezchem.dat  # phrLoadDatabaseString(frezchem.dat)
}
\description{
frezchem.dat is a database from Jonathon Toner and the
developers for the frezchem code (Toner and Sletten (2013), Marion
and coworkers). Frezchem is a Pitzer model for temperatures of 0 C
and below.
\code{\link{phrLoadDatabaseString}}.
}
\seealso{
Other Databases: 
\code{\link{Amm.dat}},
\code{\link{ColdChem.dat}},
\code{\link{Tipping_Hurley.dat}},
\code{\link{core10.dat}},
\code{\link{ex15.dat}},
\code{\link{iso.dat}},
\code{\link{llnl.dat}},
\code{\link{minteq.dat}},
\code{\link{minteq.v4.dat}},
\code{\link{phreeqc.dat}},
\code{\link{pitzer.dat}},
\code{\link{sit.dat}},
\code{\link{wateq4f.dat}}
}
\concept{Databases}
\keyword{dataset}
