% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\docType{data}
\name{iso.dat}
\alias{iso.dat}
\title{The iso.dat database.}
\source{
\url{https://www.usgs.gov/software/phreeqc-version-3/}
}
\usage{
iso.dat  # phrLoadDatabaseString(iso.dat)
}
\description{
iso.dat is a partial implementation of the individual component
approach to isotope calculations as described by Thorstenson and Parkhurst.
The database has been reformatted for use by
\code{\link{phrLoadDatabaseString}}.
}
\references{
\url{https://pubs.usgs.gov/tm/06/a43/pdf/tm6-A43.pdf}
}
\seealso{
Other Databases: 
\code{\link{Amm.dat}},
\code{\link{ColdChem.dat}},
\code{\link{Kinec.v2.dat}},
\code{\link{Kinec_v3.dat}},
\code{\link{PHREEQC_ThermoddemV1.10_15Dec2020.dat}},
\code{\link{Tipping_Hurley.dat}},
\code{\link{core10.dat}},
\code{\link{ex15.dat}},
\code{\link{frezchem.dat}},
\code{\link{llnl.dat}},
\code{\link{minteq.dat}},
\code{\link{minteq.v4.dat}},
\code{\link{phreeqc.dat}},
\code{\link{phreeqc_rates.dat}},
\code{\link{pitzer.dat}},
\code{\link{sit.dat}},
\code{\link{stimela.dat}},
\code{\link{wateq4f.dat}}
}
\concept{Databases}
\keyword{dataset}
