% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\docType{data}
\name{Kinec.v2.dat}
\alias{Kinec.v2.dat}
\title{Thermodynamic and rates database from Oelkers and coworkers.}
\source{
\url{http://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc}
}
\usage{
Kinec.v2.dat  # phrLoadDatabaseString(Kinec.v2.dat)
}
\description{
Kinec.v2.dat contains the parameters for calculating mineral
dissolution rates for primary and secondary silicate minerals using the equations
and parameters reported by Hermanska et al. (2022, 2023), and dissolution rates
for other non)-silicate mineral systems using the equations and parameters
reported by Oelkers and Addassi (2024, in preparation).
}
\references{
Hermanska et al. (2022, 2003) and Oelkers and Addassi (2024, in preparation).
}
\seealso{
Other Databases: 
\code{\link{Amm.dat}},
\code{\link{ColdChem.dat}},
\code{\link{PHREEQC_ThermoddemV1.10_15Dec2020.dat}},
\code{\link{Tipping_Hurley.dat}},
\code{\link{core10.dat}},
\code{\link{ex15.dat}},
\code{\link{frezchem.dat}},
\code{\link{iso.dat}},
\code{\link{llnl.dat}},
\code{\link{minteq.dat}},
\code{\link{minteq.v4.dat}},
\code{\link{phreeqc.dat}},
\code{\link{phreeqc_rates.dat}},
\code{\link{pitzer.dat}},
\code{\link{sit.dat}},
\code{\link{wateq4f.dat}}
}
\concept{Databases}
\keyword{dataset}
