% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{phreeqc.dat}
\alias{phreeqc.dat}
\title{The phreeqc.dat database}
\source{
\url{http://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc}
}
\usage{
phreeqc.dat  # phrLoadDatabaseString(phreeqc.dat)
}
\description{
phreeqc.dat is a phreeqc database file derived from PHREEQE,
which is consistent with wateq4f.dat, but has a smaller set of elements and
aqueous species. The database has been reformatted for use by
\code{\link{phrLoadDatabaseString}}.
}
\references{
\url{http://pubs.usgs.gov/tm/06/a43/pdf/tm6-A43.pdf}
}
\seealso{
Other Databases: \code{\link{Amm.dat}};
  \code{\link{ex15.dat}}; \code{\link{iso.dat}};
  \code{\link{llnl.dat}}; \code{\link{minteq.dat}};
  \code{\link{minteq.v4.dat}}; \code{\link{pitzer.dat}};
  \code{\link{sit.dat}}; \code{\link{wateq4f.dat}}
}
\keyword{dataset}

