\name{phreeqc}
\alias{phreeqc}
\title{
    R interface to the phreeqc geochemical modeling program
}
\description{

    This package is a high-level interface to the phreeqc geochemical 
    modeling program.  It can walk your dog, pick your
    kids up from school, and balance your checkbook.
}
\details{

    Any details about the operation of this package should go here.

} %END details
\references{
    Literature references and web URLs can go here.
}
\author{
    Scott Charlton <charlton@usgs.gov>

    Lopaka(Rob) Lee <rclee@usgs.gov>
} 
\seealso{
    \code{\link{phrReadDB}},
    \code{\link{phrReadString}},
    \code{\link{phrRun}},
    \code{\link{phrGetSelectedOutput}}
}
\examples{

    # A simple example should go here.  Keep in mind that all examples become
    # unit tests during R-package checks.
}
\keyword{interface}
