% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\docType{data}
\name{phreeqc_rates.dat}
\alias{phreeqc_rates.dat}
\title{Thermodynamic and rates database}
\source{
\url{https://www.usgs.gov/software/phreeqc-version-3/}
}
\usage{
phreeqc_rates.dat  # phrLoadDatabaseString(phreeqc_rates.dat)
}
\description{
Same as the phreeqc.dat database, but with new data blocks
RATE_PARAMETERS_HERMANSKA, RATE_PARAMETERS_PK, and
RATE_PARAMETERS_SVD that tabulate rate parameters from Hermanska
and others (2023), Palandri and Kharaka (2004), and Sverdrup and
others (2019). The Sverdrup parameters are only for two minerals
as a demonstration. Basic functions RATE_HERMANSKA, RATE_PK, and
RATE_SVD can be used to calculate rates using the corresponding
parameters.
}
\references{
Hermanska and others (2023), Palandri and Kharaka (2004),
and Sverdrup and others (2019).
}
\seealso{
Other Databases: 
\code{\link{Amm.dat}},
\code{\link{ColdChem.dat}},
\code{\link{Kinec.v2.dat}},
\code{\link{Kinec_v3.dat}},
\code{\link{PHREEQC_ThermoddemV1.10_15Dec2020.dat}},
\code{\link{Tipping_Hurley.dat}},
\code{\link{core10.dat}},
\code{\link{ex15.dat}},
\code{\link{frezchem.dat}},
\code{\link{iso.dat}},
\code{\link{llnl.dat}},
\code{\link{minteq.dat}},
\code{\link{minteq.v4.dat}},
\code{\link{phreeqc.dat}},
\code{\link{pitzer.dat}},
\code{\link{sit.dat}},
\code{\link{stimela.dat}},
\code{\link{wateq4f.dat}}
}
\concept{Databases}
\keyword{dataset}
