% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\docType{data}
\name{PHREEQC_ThermoddemV1.10_15Dec2020.dat}
\alias{PHREEQC_ThermoddemV1.10_15Dec2020.dat}
\title{Thermochemical Database from the BRGM institute (French Geological Survey)}
\source{
\url{https://www.usgs.gov/software/phreeqc-version-3/}
}
\usage{
PHREEQC_ThermoddemV1.10_15Dec2020.dat
# phrLoadDatabaseString(PHREEQC_ThermoddemV1.10_15Dec2020.dat)
}
\description{
Thermochemical Database from the BRGM institute (French Geological Survey)
}
\references{
\url{https://thermoddem.brgm.fr/}
}
\seealso{
Other Databases: 
\code{\link{Amm.dat}},
\code{\link{ColdChem.dat}},
\code{\link{Kinec.v2.dat}},
\code{\link{Kinec_v3.dat}},
\code{\link{Tipping_Hurley.dat}},
\code{\link{core10.dat}},
\code{\link{ex15.dat}},
\code{\link{frezchem.dat}},
\code{\link{iso.dat}},
\code{\link{llnl.dat}},
\code{\link{minteq.dat}},
\code{\link{minteq.v4.dat}},
\code{\link{phreeqc.dat}},
\code{\link{phreeqc_rates.dat}},
\code{\link{pitzer.dat}},
\code{\link{sit.dat}},
\code{\link{wateq4f.dat}}
}
\concept{Databases}
