% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetDumpFileOn}
\alias{phrGetDumpFileOn}
\title{Retrieve the current value of the dump file switch.}
\usage{
phrGetDumpFileOn()
}
\value{
TRUE if errors are currently being written to file.
}
\description{
Retrieve the current value of the dump file switch.
}
\references{
\url{ftp://brrftp.cr.usgs.gov/pub/charlton/iphreeqc/IPhreeqc.pdf}
}
\seealso{
Other Dump: \code{\link{phrGetDumpFileName}};
  \code{\link{phrGetDumpStringsOn}};
  \code{\link{phrGetDumpStrings}};
  \code{\link{phrSetDumpFileName}};
  \code{\link{phrSetDumpFileOn}};
  \code{\link{phrSetDumpStringsOn}}
}

