% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetErrorStringsOn}
\alias{phrGetErrorStringsOn}
\title{Retrieve the current value of the error strings switch.}
\usage{
phrGetErrorStringsOn()
}
\value{
TRUE if errors are currently being written to file.
}
\description{
Retrieve the current value of the error strings switch.
}
\references{
\url{ftp://brrftp.cr.usgs.gov/pub/charlton/iphreeqc/IPhreeqc.pdf}
}
\seealso{
Other Error: \code{\link{phrGetErrorFileName}},
  \code{\link{phrGetErrorFileOn}},
  \code{\link{phrGetErrorStrings}},
  \code{\link{phrSetErrorFileName}},
  \code{\link{phrSetErrorFileOn}},
  \code{\link{phrSetErrorStringsOn}}
}
