% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetLogStringsOn}
\alias{phrGetLogStringsOn}
\title{Retrieve the current value of the log strings switch.}
\usage{
phrGetLogStringsOn()
}
\value{
TRUE if log messages are currently being stored as a character vector.
}
\description{
Retrieve the current value of the log strings switch.
}
\references{
\url{https://www.usgs.gov/software/phreeqc-version-3/}
}
\seealso{
Other Log: 
\code{\link{phrGetLogFileName}()},
\code{\link{phrGetLogFileOn}()},
\code{\link{phrGetLogStrings}()},
\code{\link{phrSetLogFileName}()},
\code{\link{phrSetLogFileOn}()},
\code{\link{phrSetLogStringsOn}()}
}
\concept{Log}
