% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phrGetOutputFileOn}
\alias{phrGetOutputFileOn}
\title{Retrieve the current value of the output file switch.}
\usage{
phrGetOutputFileOn()
}
\value{
TRUE if errors are currently being written to file.
}
\description{
Retrieve the current value of the output file switch.
}
\references{
\url{ftp://brrftp.cr.usgs.gov/pub/charlton/iphreeqc/IPhreeqc.pdf}
}
\seealso{
Other Output: \code{\link{phrGetOutputFileName}};
  \code{\link{phrGetOutputStringsOn}};
  \code{\link{phrGetOutputStrings}};
  \code{\link{phrSetOutputFileName}};
  \code{\link{phrSetOutputFileOn}};
  \code{\link{phrSetOutputStringsOn}}
}

