\name{phrReadString}
\alias{phrReadString}
\title{
    Read phreeqc input-file commands
}
\description{

    phrReadString reads phreeqc input-file commands ...
}
\synopsis{
    phrReadString(string)
}
\usage{
    phrReadString(string)
}
\arguments{
\item{string}{
    A character string containing phreeqc input-file commands.
} 
} %END arguments
\details{

    Any details about the operation of this function should go here.

} %END details
\value{
    This function returns NULL.
}
\references{
    Literature references and web URLs can go here.
}
\author{
    Scott Charlton <charlton@usgs.gov>

    Lopaka(Rob) Lee <rclee@usgs.gov>
} 
\seealso{
    \code{\link{phrReadDB}},
    \code{\link{phrRun}},
    \code{\link{phrGetSelectedOutput}}
}

\examples{

    # Simple example here
}
\keyword{interface}
