\name{phrRun}
\alias{phrRun}
\title{
    Run phreeqc
}
\description{

    phrRun executes a phreeqc run
}
\synopsis{
    phrRun()
}
\usage{
    phrRun()
}
\arguments{
} %END arguments
\details{

    Any details about the operation of this function should go here.
} %END details
\value{
    This function returns NULL.
}
\references{
    Literature references and web URLs can go here.
}
\author{
    Scott Charlton <charlton@usgs.gov>

    Lopaka(Rob) Lee <rclee@usgs.gov>
} 
\seealso{
    \code{\link{phrReadString}},
    \code{\link{phrRunFile}},
    \code{\link{phrGetSelectedOutput}}
}

\examples{

    #phrRun()
}
\keyword{interface}
