% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\docType{data}
\name{stimela.dat}
\alias{stimela.dat}
\title{The stimela.dat database.}
\usage{
stimela.dat  # phrLoadDatabaseString(stimela.dat)
}
\description{
stimela.dat is a database for use in drinking-water and waste-water
treatment from Peter de Moel and Omnisys. The database has been reformatted
for use by \code{\link{phrLoadDatabaseString}}.
}
\seealso{
Other Databases: 
\code{\link{Amm.dat}},
\code{\link{ColdChem.dat}},
\code{\link{Kinec.v2.dat}},
\code{\link{Kinec_v3.dat}},
\code{\link{PHREEQC_ThermoddemV1.10_15Dec2020.dat}},
\code{\link{Tipping_Hurley.dat}},
\code{\link{core10.dat}},
\code{\link{ex15.dat}},
\code{\link{frezchem.dat}},
\code{\link{iso.dat}},
\code{\link{llnl.dat}},
\code{\link{minteq.dat}},
\code{\link{minteq.v4.dat}},
\code{\link{phreeqc.dat}},
\code{\link{phreeqc_rates.dat}},
\code{\link{pitzer.dat}},
\code{\link{sit.dat}},
\code{\link{wateq4f.dat}}
}
\concept{Databases}
\keyword{dataset}
