% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\docType{data}
\name{Tipping_Hurley.dat}
\alias{Tipping_Hurley.dat}
\title{The Tipping_Hurley.dat database}
\usage{
Tipping_Hurley.dat  # phrLoadDatabaseString(Tipping_Hurley.dat)
}
\description{
Tipping_Hurley.dat is a database for organic-ligand
binding approximating WHAM by Tipping and Hurley.
\code{\link{phrLoadDatabaseString}}.
}
\seealso{
Other Databases: 
\code{\link{Amm.dat}},
\code{\link{ColdChem.dat}},
\code{\link{core10.dat}},
\code{\link{ex15.dat}},
\code{\link{frezchem.dat}},
\code{\link{iso.dat}},
\code{\link{llnl.dat}},
\code{\link{minteq.dat}},
\code{\link{minteq.v4.dat}},
\code{\link{phreeqc.dat}},
\code{\link{pitzer.dat}},
\code{\link{sit.dat}},
\code{\link{wateq4f.dat}}
}
\concept{Databases}
\keyword{dataset}
