% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\docType{data}
\name{wateq4f.dat}
\alias{wateq4f.dat}
\title{The wateq4f.dat database.}
\source{
\url{https://www.usgs.gov/software/phreeqc-version-3/}
}
\usage{
wateq4f.dat  # phrLoadDatabaseString(wateq4f.dat)
}
\description{
wateq4f.dat is a database derived from WATEQ4F. The database
has been reformatted for use by \code{\link{phrLoadDatabaseString}}.
}
\references{
\url{https://pubs.usgs.gov/tm/06/a43/pdf/tm6-A43.pdf}
}
\seealso{
Other Databases: 
\code{\link{Amm.dat}},
\code{\link{ColdChem.dat}},
\code{\link{Kinec.v2.dat}},
\code{\link{Kinec_v3.dat}},
\code{\link{PHREEQC_ThermoddemV1.10_15Dec2020.dat}},
\code{\link{Tipping_Hurley.dat}},
\code{\link{core10.dat}},
\code{\link{ex15.dat}},
\code{\link{frezchem.dat}},
\code{\link{iso.dat}},
\code{\link{llnl.dat}},
\code{\link{minteq.dat}},
\code{\link{minteq.v4.dat}},
\code{\link{phreeqc.dat}},
\code{\link{phreeqc_rates.dat}},
\code{\link{pitzer.dat}},
\code{\link{sit.dat}},
\code{\link{stimela.dat}}
}
\concept{Databases}
\keyword{dataset}
