% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\docType{data}
\name{ColdChem.dat}
\alias{ColdChem.dat}
\title{The ColdChem.dat database.}
\source{
\url{http://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc}
}
\usage{
ColdChem.dat  # phrLoadDatabaseString(ColdChem.dat)
}
\description{
ColdChem.dat is a low-temperature thermodynamic model for the
Na-K-Ca-Mg-Cl system incorporating new experimental heat capacities in
KCl, MgCl2, and CaCl2 solutions. Jonathan D. Toner and David C. Catling
The database has been reformatted for use by \code{\link{phrLoadDatabaseString}}.
}
\seealso{
Other Databases: 
\code{\link{Amm.dat}},
\code{\link{Tipping_Hurley.dat}},
\code{\link{core10.dat}},
\code{\link{ex15.dat}},
\code{\link{frezchem.dat}},
\code{\link{iso.dat}},
\code{\link{llnl.dat}},
\code{\link{minteq.dat}},
\code{\link{minteq.v4.dat}},
\code{\link{phreeqc.dat}},
\code{\link{pitzer.dat}},
\code{\link{sit.dat}},
\code{\link{wateq4f.dat}}
}
\concept{Databases}
\keyword{dataset}
