\docType{data}
\name{minteq.v4.dat}
\alias{minteq.v4.dat}
\title{The minteq.v4.dat database.}
\source{
\url{http://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc}
}
\usage{
minteq.v4.dat  # phrLoadDatabaseString(minteq.v4.dat)
}
\description{
minteq.v4.dat is a database derived from MINTEQA2 version
4. The database has been reformatted for use by
\code{\link{phrLoadDatabaseString}}.
}
\references{
\url{http://pubs.usgs.gov/tm/06/a43/pdf/tm6-A43.pdf}
}
\seealso{
Other Databases: \code{\link{Amm.dat}};
\code{\link{ex15.dat}}; \code{\link{iso.dat}};
\code{\link{llnl.dat}}; \code{\link{minteq.dat}};
\code{\link{phreeqc.dat}}; \code{\link{pitzer.dat}};
\code{\link{sit.dat}}; \code{\link{wateq4f.dat}}
}
\keyword{dataset}

