% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetAccumulatedLines}
\alias{phrGetAccumulatedLines}
\title{Retrieve the accumulated input.}
\usage{
phrGetAccumulatedLines()
}
\value{
A character vector containing the accumulated input.
}
\description{
Returns the accumulated input as a character vector.
}
\examples{

# This example loads some keyword input and displays the contents.
phrAccumulateLine("SOLUTION 1")
phrAccumulateLine("END")
cat("The accumulated input is:", phrGetAccumulatedLines(), sep = "\n")

}
\references{
\url{https://water.usgs.gov/water-resources/software/PHREEQC/IPhreeqc.pdf}
}
\seealso{
Other Accumulate: 
\code{\link{phrAccumulateLine}()},
\code{\link{phrClearAccumulatedLines}()},
\code{\link{phrRunAccumulated}()}
}
\concept{Accumulate}
