% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetDumpFileOn}
\alias{phrGetDumpFileOn}
\title{Retrieve the current value of the dump file switch.}
\usage{
phrGetDumpFileOn()
}
\value{
TRUE if output as defined by the DUMP data block are currently being written to file.
}
\description{
Retrieve the current value of the dump file switch.
}
\references{
\url{https://water.usgs.gov/water-resources/software/PHREEQC/IPhreeqc.pdf}
}
\seealso{
Other Dump: 
\code{\link{phrGetDumpFileName}()},
\code{\link{phrGetDumpStringsOn}()},
\code{\link{phrGetDumpStrings}()},
\code{\link{phrSetDumpFileName}()},
\code{\link{phrSetDumpFileOn}()},
\code{\link{phrSetDumpStringsOn}()}
}
\concept{Dump}
