% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetDumpStringsOn}
\alias{phrGetDumpStringsOn}
\title{Retrieve the current value of the dump strings switch.}
\usage{
phrGetDumpStringsOn()
}
\value{
TRUE if output as defined by the DUMP data block are currently being stored as a character vector.
}
\description{
Retrieve the current value of the dump strings switch.
}
\references{
\url{https://water.usgs.gov/water-resources/software/PHREEQC/IPhreeqc.pdf}
}
\seealso{
Other Dump: 
\code{\link{phrGetDumpFileName}()},
\code{\link{phrGetDumpFileOn}()},
\code{\link{phrGetDumpStrings}()},
\code{\link{phrSetDumpFileName}()},
\code{\link{phrSetDumpFileOn}()},
\code{\link{phrSetDumpStringsOn}()}
}
\concept{Dump}
