% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetErrorFileOn}
\alias{phrGetErrorFileOn}
\title{Retrieve the current value of the error file switch.}
\usage{
phrGetErrorFileOn()
}
\value{
TRUE if error messages are currently being written to file.
}
\description{
Retrieve the current value of the error file switch.
}
\references{
\url{https://water.usgs.gov/water-resources/software/PHREEQC/IPhreeqc.pdf}
}
\seealso{
Other Error: 
\code{\link{phrGetErrorFileName}()},
\code{\link{phrGetErrorStringsOn}()},
\code{\link{phrGetErrorStrings}()},
\code{\link{phrSetErrorFileName}()},
\code{\link{phrSetErrorFileOn}()},
\code{\link{phrSetErrorStringsOn}()}
}
\concept{Error}
