% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetErrorStringsOn}
\alias{phrGetErrorStringsOn}
\title{Retrieve the current value of the error strings switch.}
\usage{
phrGetErrorStringsOn()
}
\value{
TRUE if error messages are currently being stored as a character vector.
}
\description{
Retrieve the current value of the error strings switch.
}
\references{
\url{https://water.usgs.gov/water-resources/software/PHREEQC/IPhreeqc.pdf}
}
\seealso{
Other Error: 
\code{\link{phrGetErrorFileName}()},
\code{\link{phrGetErrorFileOn}()},
\code{\link{phrGetErrorStrings}()},
\code{\link{phrSetErrorFileName}()},
\code{\link{phrSetErrorFileOn}()},
\code{\link{phrSetErrorStringsOn}()}
}
\concept{Error}
