% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phrGetLogFileOn}
\alias{phrGetLogFileOn}
\title{Retrieve the current value of the log file switch.}
\usage{
phrGetLogFileOn()
}
\value{
TRUE if errors are currently being written to file.
}
\description{
Retrieve the current value of the log file switch.
}
\references{
\url{ftp://brrftp.cr.usgs.gov/pub/charlton/iphreeqc/IPhreeqc.pdf}
}
\seealso{
Other Log: \code{\link{phrGetLogFileName}};
  \code{\link{phrGetLogStringsOn}};
  \code{\link{phrGetLogStrings}};
  \code{\link{phrSetLogFileName}};
  \code{\link{phrSetLogFileOn}};
  \code{\link{phrSetLogStringsOn}}
}

