% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetLogFileOn}
\alias{phrGetLogFileOn}
\title{Retrieve the current value of the log file switch.}
\usage{
phrGetLogFileOn()
}
\value{
TRUE if log messages are currently being written to file.
}
\description{
Retrieve the current value of the log file switch.
}
\references{
\url{https://water.usgs.gov/water-resources/software/PHREEQC/IPhreeqc.pdf}
}
\seealso{
Other Log: 
\code{\link{phrGetLogFileName}()},
\code{\link{phrGetLogStringsOn}()},
\code{\link{phrGetLogStrings}()},
\code{\link{phrSetLogFileName}()},
\code{\link{phrSetLogFileOn}()},
\code{\link{phrSetLogStringsOn}()}
}
\concept{Log}
