% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phrGetLogStringsOn}
\alias{phrGetLogStringsOn}
\title{Retrieve the current value of the log strings switch.}
\usage{
phrGetLogStringsOn()
}
\value{
TRUE if errors are currently being written to file.
}
\description{
Retrieve the current value of the log strings switch.
}
\references{
\url{http://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc}
}
\seealso{
Other Log: \code{\link{phrGetLogFileName}};
  \code{\link{phrGetLogFileOn}};
  \code{\link{phrGetLogStrings}};
  \code{\link{phrSetLogFileName}};
  \code{\link{phrSetLogFileOn}};
  \code{\link{phrSetLogStringsOn}}
}

