% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetOutputFileOn}
\alias{phrGetOutputFileOn}
\title{Retrieve the current value of the output file switch.}
\usage{
phrGetOutputFileOn()
}
\value{
TRUE if output messages are currently being written to file.
}
\description{
Retrieve the current value of the output file switch.
}
\references{
\url{https://water.usgs.gov/water-resources/software/PHREEQC/IPhreeqc.pdf}
}
\seealso{
Other Output: 
\code{\link{phrGetOutputFileName}()},
\code{\link{phrGetOutputStringsOn}()},
\code{\link{phrGetOutputStrings}()},
\code{\link{phrSetOutputFileName}()},
\code{\link{phrSetOutputFileOn}()},
\code{\link{phrSetOutputStringsOn}()}
}
\concept{Output}
