% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetOutputStringsOn}
\alias{phrGetOutputStringsOn}
\title{Retrieve the current value of the output strings switch.}
\usage{
phrGetOutputStringsOn()
}
\value{
TRUE if errors are currently being written to file.
}
\description{
Retrieve the current value of the output strings switch.
}
\references{
\url{ftp://brrftp.cr.usgs.gov/pub/charlton/iphreeqc/IPhreeqc.pdf}
}
\seealso{
Other Output: \code{\link{phrGetOutputFileName}};
  \code{\link{phrGetOutputFileOn}};
  \code{\link{phrGetOutputStrings}};
  \code{\link{phrSetOutputFileName}};
  \code{\link{phrSetOutputFileOn}};
  \code{\link{phrSetOutputStringsOn}}
}

