% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc.R
\name{phrGetOutputStringsOn}
\alias{phrGetOutputStringsOn}
\title{Retrieve the current value of the output strings switch.}
\usage{
phrGetOutputStringsOn()
}
\value{
TRUE if output messages are currently being stored as a character vector.
}
\description{
Retrieve the current value of the output strings switch.
}
\references{
\url{https://water.usgs.gov/water-resources/software/PHREEQC/IPhreeqc.pdf}
}
\seealso{
Other Output: 
\code{\link{phrGetOutputFileName}()},
\code{\link{phrGetOutputFileOn}()},
\code{\link{phrGetOutputStrings}()},
\code{\link{phrSetOutputFileName}()},
\code{\link{phrSetOutputFileOn}()},
\code{\link{phrSetOutputStringsOn}()}
}
\concept{Output}
