\name{phrGetSelectedOutput}
\alias{phrGetSelectedOutput}
\title{
    Return the results of a phreeqc run.
}
\description{
    phrGetSelectedOutput return the results of a phreeqc run specified
    in the selected output block of a phreeqc input file.

}
\synopsis{
    phrGetSelectedOutput()
}
\usage{
    phrGetSelectedOutput()
}
\arguments{
} %END arguments
\details{

    Any details about the operation of this function should go here.

} %END details
\value{
    Returns a data frame containing ... 
}
\references{
    Literature references and web URLs can go here.
}
\author{
    Scott Charlton <charlton@usgs.gov>

    Lopaka(Rob) Lee <rclee@usgs.gov>
} 
\seealso{
    \code{\link{phrReadDB}},
    \code{\link{phrReadString}},
    \code{\link{phrRun}}
}
\examples{

    # Simple example here
}
\keyword{interface}
